Give div containers to connect to
this HTML commit will give me some div containers
to connect to because I am having a hard time
conceptualizing how to build things with out a "thing"
to attach things to.

also I have been playing with the window.prompt function
for this project. so that will show up in the .js file.

Thank You.